+++ app/src/main/java/com/kabouzeid/gramophone/util/MusicUtil.java
@@ -182,7 +182,7 @@ public class MusicUtil {
         return trackNumberToFix % 1000;
     }
 
-    public static void insertAlbumArt(@NonNull Context context, int albumId, String path) {
+    public static void insertAlbumArt(@NonNull Context context, long albumId, String path) {
         ContentResolver contentResolver = context.getContentResolver();
 
         Uri artworkUri = Uri.parse("content://media/external/audio/albumart");
@@ -196,7 +196,7 @@ public class MusicUtil {
         contentResolver.notifyChange(artworkUri, null);
     }
 
-    public static void deleteAlbumArt(@NonNull Context context, int albumId) {
+    public static void deleteAlbumArt(@NonNull Context context, long albumId) {
         ContentResolver contentResolver = context.getContentResolver();
         Uri localUri = Uri.parse("content://media/external/audio/albumart");
         contentResolver.delete(ContentUris.withAppendedId(localUri, albumId), null, null);
@@ -248,7 +248,7 @@ public class MusicUtil {
                 // as from the album art cache
                 cursor.moveToFirst();
                 while (!cursor.isAfterLast()) {
-                    final int id = cursor.getInt(0);
+                    final long id = cursor.getLong(0);
                     final Song song = SongLoader.getSong(context, id);
                     MusicPlayerRemote.removeFromQueue(song);
                     cursor.moveToNext();
@@ -258,7 +258,7 @@ public class MusicUtil {
                 // Step 2: Remove files from card
                 cursor.moveToFirst();
                 while (!cursor.isAfterLast()) {
-                    final int id = cursor.getInt(0);
+                    final long id = cursor.getLong(0);
                     final String name = cursor.getString(1);
                     try { // File.delete can throw a security exception
                         final File f = new File(name);
